requirements.txt
amqp==2.6.0
asgiref==3.2.7
astroid==2.4.1
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.6
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
Django==3.0.6
django-bootstrap-modal-forms==2.0.0
django-cors-headers==3.3.0
django-crispy-forms==1.9.2
django-materialize==1.0.1a2
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-rest-knox==4.1.0
django-shapeshifter==18.9.23
django-tables2==2.3.1
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
future==0.18.2
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.2
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
vine==1.3.0
wrapt==1.12.1